authorization

+++ app/Http/Controllers/NewsCategoryController.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
-
+use Sentinel;
 use Illuminate\Http\Request;
 use Illuminate\Database\QueryException;
 use App\Repositories\NewsCategory\NewsCategoryContract;
@@ -14,15 +14,29 @@ class NewsCategoryController extends Controller
     }
 
     public function index() {
+      if(!Sentinel::check()){
+
+      }else{
+
+      }
       $news_categories = $this->repo->findAll();
       return view('news_category.index')->with('news_categories', $news_categories);
     }
     
     public function create() {
-		return view('news_category.create');
+      if(!Sentinel::check()){
+          return redirect()->route('auth.login.get');
+      }else{
+        return view('news_category.create');
+      }
+		
     }
     
     public function store(Request $request) {
+      if(!Sentinel::check()){
+        return redirect()->route('auth.login.get');
+      }
+      else{
       $this->validate($request, [
         'name' => 'required',
         'description' => 'required',
@@ -54,19 +68,32 @@ class NewsCategoryController extends Controller
         }
       }
     }
+   }
 
     public function show($slug) {
-      $news_category = $this->repo->findBySlug($slug);
+      if(!Sentinel::check()){
+        return redirect()->route('auth.login.get');
+      }else{
+       $news_category = $this->repo->findBySlug($slug);
       return view('news_category.show')->with('news_category', $news_category);
+      }
     }
 
     public function edit($slug) {
-      $news_category = $this->repo->findBySlug($slug);
-      return view('news_category.edit')->with('news_category', $news_category);
+      if(!Sentinel::check()){
+        return redirect()->route('auth.login.get');
+      }
+      else{
+        $news_category = $this->repo->findBySlug($slug);
+        return view('news_category.edit')->with('news_category', $news_category);
+      }
     }
 
     public function update(Request $request, $slug) {
-      $news_category = $this->repo->update($request, $slug);
+      if(!Sentinel::check()){
+        return redirect()->route('auth.login.get');
+      }else{
+        $news_category = $this->repo->update($request, $slug);
       $notification = array(
         'message' => "News Category $news_category->name updated successfully",
         'alert-type' => 'success'
@@ -75,22 +102,28 @@ class NewsCategoryController extends Controller
       if($news_category->id) {
         return redirect()->route('newsCategory.index')->with($notification);
       }
+      }
+      
     }
 
     public function delete($slug) {
-		
-      if ($this->repo->remove($slug)) {
-        $notification = array(
-          'message' => "News Category deleted successfully",
-          'alert-type' => 'success'
-        );
-        return redirect()->back()->with($notification);
-       } else {
-        $error = array(
-          'message' => 'Error Deleting News Category',
-          'alert-type' => 'error'
-        );
-        return back()->with($error);
-      }    
-    }
+      if(!Sentinel::check()){
+        return redirect()->route('auth.login.get');
+      }else{
+        if ($this->repo->remove($slug)) {
+          $notification = array(
+            'message' => "News Category deleted successfully",
+            'alert-type' => 'success'
+          );
+          return redirect()->back()->with($notification);
+         } else {
+          $error = array(
+            'message' => 'Error Deleting News Category',
+            'alert-type' => 'error'
+          );
+          return back()->with($error);
+        }    
+      }
+      }
+     
 }
